Refactoring the code in components
